pyproject.toml
[build-system]
requires = [
    "setuptools",
    "robotframework",
    "robotframework-seleniumlibrary"
    ]
build-backend = "setuptools.build_meta"

[project]
name = "robotframework-DEMOQA-driver"
authors = [{name="Artur Ziolkowksi", email="artur.k.ziolkowski@gmail.com"},
]
description = "Database Library for Robot Framework"
readme = "README.md"
requires-python = ">=3.6"
dependencies = [
    "robotframework",
    "robotframework-seleniumlibrary"
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
license = {text = "Apache License 2.0"}
dynamic = ["version"]

[project.urls]
"Homepage" = "https://github.com/ajadach/robotframework-DEMOQA-driver"
"Keyword docs" = "http://marketsquare.github.io/robotframework-DEMOQA-driver/"

[tool.setuptools.dynamic]
version = {attr = "DatabaseLibrary.__version__"}

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120